refactor: 홈 화면 랭킹 관련 컴포넌트 및 모델 패키지 구조 개선
- 랭킹 관련 클래스들을 `com.yagubogu.ui.home.ranking` 하위 패키지로 이동
    - 모델: `RankingItem`, `RankingProfileItem`
    - 컴포넌트: `Ranking.kt`
- `Ranking.kt` 내부에 선언되어 있던 `RankingMemberProfile` 및 `RankMedal`을 별도 파일(`RankingMemberProfile.kt`)로 분리 및 모듈화
- 패키지 위치 변경에 따른 `HomeViewModel`, `HomeScreen`, `StatsUiMapper` 등 관련 참조 경로 수정
