+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/home/ranking/model/RankingItem.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.ui.home.model
+package com.yagubogu.ui.home.ranking.model
 
 sealed interface RankingItem {
     val topRankings: List<RankingProfileItem>
